adding missing update
