+++ src/assets/styles/_variables.scss
@@ -9,7 +9,6 @@ $black: #444;
 $body-color: RGB(240, 248, 255);
 $body-color: white;
 $grey: grey;
-$faded-grey: rgba(192, 192, 192, 0.5);
 $item-height: 62px;
 $item-margin: 14px;
 $error-color: #6b0103;
